% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_effects_ieu.R
\name{align_effects_ieu}
\alias{align_effects_ieu}
\title{Align alleles looked up from IEU Open GWAS using ieugwasr package}
\usage{
align_effects_ieu(phe, ref, ref_snp_col = "snp", ref_ea = "A1", ref_nea = "A2")
}
\arguments{
\item{phe}{Output of ieugwasr::phewas or ieugwasr::associations}

\item{ref}{Reference data frame}

\item{ref_ea}{Name of effect allele column in ref}

\item{ref_nea}{Name of non effect allele column in ref}
}
\description{
Align alleles looked up from IEU Open GWAS using ieugwasr package
}
