% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R.R
\name{condition}
\alias{condition}
\title{Project matrix to nearest well conditioned positive definite matrix.}
\usage{
condition(R, cond_num = 1e+05)
}
\arguments{
\item{R}{Matrix}

\item{cond_num}{Maximum allowable condition number (max(eigenvalue)/min(eigenvalue))}
}
\description{
Project matrix to nearest well conditioned positive definite matrix.
}
