% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_L.R
\name{est_L}
\alias{est_L}
\title{Estimate L}
\usage{
est_L(B_hat, S_hat, R, N, tau, fit, adjust = TRUE, tol = 1e-15)
}
\arguments{
\item{B_hat}{matrix of effect estimates}

\item{S_hat}{matrix of standard errors}

\item{R}{Estimated residual correlation of rows of B_hat}

\item{N}{Sample size}

\item{tau}{vector length n traits of per-trait precision (1/variance) of theta}

\item{fit}{Flash fit}

\item{adjust.}{Whether to adjust estimates to use standardized effects.}
}
\description{
Estimate L
}
