% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ff.R, R/fit_ff2.R
\name{fit_ff}
\alias{fit_ff}
\title{Fit with fixed factors}
\usage{
fit_ff(B_hat, S_hat, N, R, kmax = 100, zero_thresh = 1e-15, adjust = TRUE)

fit_ff(B_hat, S_hat, N, R, kmax = 100, zero_thresh = 1e-15, adjust = TRUE)
}
\arguments{
\item{B_hat}{Matrix of (non-standardized) effect estimates SNPs by traits}

\item{S_hat}{Matrix of standard errors SNPs by traits}

\item{N}{Vector of sample sizes length equal to number of traits}

\item{R}{Estimated residual correlation matrix}

\item{kmax}{Maximum number of factors}

\item{zero_thresh}{Threshold for setting eigenvalues of R to zero}
}
\value{
A list with elements fit, B_hat, L_hat, F_hat

A list with elements fit, B_hat, L_hat, F_hat
}
\description{
Fit with fixed factors

Fit with fixed factors
}
