% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ff_rewrite.R
\name{fit_ff_new}
\alias{fit_ff_new}
\title{Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects.}
\usage{
fit_ff_new(
  Z_hat,
  B_std,
  N,
  R,
  zero_thresh = 1e-15,
  method = "sequential",
  max_ev_percent = 1,
  S_NULL = FALSE
)
}
\arguments{
\item{Z_hat}{A matrix of z-scores}

\item{B_std}{A matrix of standardized effects. Supply only one of B_std or Z_hat}

\item{N}{Vector of sample sizes length equal to number of traits. N is only
needed if using B_std.}

\item{R}{Estimated residual correlation matrix}

\item{zero_thresh}{Threshold for setting eigenvalues of R to zero}

\item{S_NULL}{If TRUE, will fit EBMF without supplying S.}

\item{kmax}{Maximum number of factors}

\item{max_ev_percent.}{Only the eigenvectors of R explaining the top percentage of
variance are retained. Defaults to 1, retaining all eigenvectors.}
}
\value{
A list with elements fit, B_hat, L_hat, F_hat
}
\description{
Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects.
}
