% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ff_future.R
\name{fit_ff_prefit}
\alias{fit_ff_prefit}
\title{Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.}
\usage{
fit_ff_prefit(
  Z_hat,
  B_std,
  N,
  R,
  kmax,
  zero_thresh = 1e-15,
  method = "extrapolate",
  max_ev_percent = 1,
  S_inf = c(10, 1),
  num_prefits = 1,
  max_prefit_iter = 500,
  min_var_ratio = 1
)
}
\arguments{
\item{Z_hat}{A matrix of z-scores}

\item{B_std}{A matrix of standardized effects. Supply only one of B_std or Z_hat}

\item{N}{Vector of sample sizes length equal to number of traits. N is only
needed if using B_std.}

\item{R}{Estimated residual correlation matrix}

\item{kmax}{Maximum number of factors. Defaults to 2*ntraits for plain or ntraits
if using the fixed factor correction.}

\item{zero_thresh}{Threshold for setting eigenvalues of R to zero}

\item{S_inf}{A vector supplying a series of variance inflation factors for fitting ebmf.
S_inf should be decreasing and the last element should be 1.}

\item{max_ev_percent.}{Only the eigenvectors of R explaining the top percentage of
variance are retained. Defaults to 1, retaining all eigenvectors.}
}
\value{
A list with elements fit, Y, L_hat, F_hat
}
\description{
Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.
}
