% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_plain.R
\name{fit_plain}
\alias{fit_plain}
\title{Fit plain vanilla FLASH to standardized effects}
\usage{
fit_plain(B_hat, S_hat, N, kmax = 100)
}
\arguments{
\item{B_hat}{Matrix of (non-standardized) effect estimates SNPs by traits}

\item{S_hat}{Matrix of standard errors SNPs by traits}

\item{N}{Vector of sample sizes length equal to number of traits}

\item{kmax}{Maximum number of factors}
}
\value{
A list with elements fit, B_hat, L_hat, F_hat
}
\description{
Fit plain vanilla FLASH to standardized effects
}
