% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_to_pathway.R
\name{genes_within}
\alias{genes_within}
\title{Get genes with TSS within a distance of variants}
\usage{
genes_within(anno_df, gene_df, dist = 50000)
}
\description{
Get genes with TSS within a distance of variants
}
