% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_to_pathway.R
\name{get_eqtl_genes}
\alias{get_eqtl_genes}
\title{Extract eQTLs from GTEx data}
\usage{
get_eqtl_genes(anno_df, gtex_files, tissue_names)
}
\arguments{
\item{anno_df}{is produced by annotate_snp_loc}

\item{gtex_files}{file list. Should be a list of tsv files with columns chr, pos, and gene_id}
}
\description{
You can find files in /project2/compbio/external_public_supp/GTEx/V8/GTEx_Analysis_v8_eQTL/*chrpos.tsv.gz
}
