% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfa_intervals.R
\name{gfa_credints}
\alias{gfa_credints}
\title{GFA Credible Intervals}
\usage{
gfa_credints(
  gfa_fit,
  nsamp = 500,
  level = 0.05,
  type = c("pve", "F", "L"),
  variant_ix = NULL
)
}
\arguments{
\item{gfa_fit}{An object output by gfa_fit.}

\item{nsamp}{Number of posterior samples.}

\item{level}{Level for credible intervals (e.g. 0.05 results in 95\% credible intervals).}

\item{type}{A vector including some or all elements "pve", "F", or "L", see details.}

\item{variant_ix}{Optional list of indices if type includes "L".}
}
\description{
Compute credible intervals for various estimated parameters.
}
\details{
This function will compute credible intervals for the PVE matrix, F, or L depending on the
choice in "type". The returned object will include <type>\_lower and <type>\_upper for
each element in type.

Please note that computing credible intervals for all elements of L can be very
time consuming, since L is typically very large. Your needs may be better served
by computing frequentist estimates and p-values for L using gfa_loadings_gls().
}
