% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfa_fit.R
\name{gfa_fit}
\alias{gfa_fit}
\title{Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.}
\usage{
gfa_fit(Z_hat, B_std, N, R, params = gfa_default_parameters())
}
\arguments{
\item{Z_hat}{A matrix of z-scores}

\item{B_std}{A matrix of standardized effects. Supply only one of B_std or Z_hat}

\item{N}{Vector of sample sizes length equal to number of traits. N is only
needed if using B_std.}

\item{R}{Estimated residual correlation matrix}

\item{params}{List of parameters. For most users this can be left at the default values. See details.}
}
\value{
A list with elements fit, Y, L_hat, F_hat
}
\description{
Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.
}
\details{
The params list includes:

kmax, the maximum number of factors

max_iter: maximum number of iterations

min_ev, max_lr_percent, lr_zero_thresh: We are approximating the matrix R as VDV^T + lambda I
where lambda is the smallest eigenvalue of R. min_ev specifies the smallest acceptable value of lambda.
max_lr_percent specifies the proportion of variance of (R - lambda I) that will be retained. This defaults
to 1 making the approximation exact. Eigenvalues of (R-lambda I) that are less than lr_zero_thresh will
be set to zero.

ridge_penalty

extrapolate: value of extrapolate passed to flash_backfit

ebnm_fn_F, ebnm_fn_L

init_fn

fixed_truncate

duplicate_check_thresh
}
