% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_rg.R
\name{ldsc_rg}
\alias{ldsc_rg}
\title{Cross-Trait LD score regression}
\usage{
ldsc_rg(
  ld_score,
  ld_size,
  z1,
  z2,
  sample_size_1,
  sample_size_2,
  blocks = NULL,
  h2_1 = NULL,
  h2_2 = NULL,
  intercept = NULL,
  intercept_h2_1 = NULL,
  intercept_h2_2 = NULL,
  step1_chisq_max = 30,
  chi2_thr2 = Inf,
  ncores = 1
)
}
\arguments{
\item{ld_score}{Vector of LD scores.}

\item{ld_size}{Number of variants used to compute `ld_score`.}

\item{z1}{Vector of z-scores for trait 1.}

\item{z2}{Vector of z-scores for trait 2.}

\item{sample_size_1}{Sample size of GWAS for trait 1.
Possibly a vector, or just a single value.}

\item{sample_size_2}{Sample size of GWAS for trait 2.
Possibly a vector, or just a single value.}

\item{blocks}{Either a single number specifying the number of blocks,
or a vector of integers specifying the block number of each `chi2` value.
Default is `200` for `snp_ldsc()`, dividing into 200 blocks of approximately
equal size. `NULL` can also be used to skip estimating standard errors,
which is the default for `snp_ldsc2()`.}

\item{intercept}{You can constrain the intercept to some value (e.g. 0).
Default is `NULL` (the intercept is estimated).
Use a value of 0 if you are sure there is no overlap between GWAS samples.}

\item{intercept_h2_1}{Intercept for heritability of trait 1 (default is NULL so the intercept is estimated).}

\item{intercept_h2_2}{Intercept for heritability of trait 2 (default is NULL so the intercept is estimated).}

\item{step1_chisq_max}{Threshold on `chi2` in step 1. Default is `30`.}

\item{chi2_thr2}{Threshold on `chi2` in step 2. Default is `Inf` (none).}
}
\value{
Vector of 4 values (or only the first 2 if `blocks = NULL`):
 - `[["int"]]`: LDSC regression intercept,
 - `[["int_se"]]`: SE of this intercept,
 - `[["h2"]]`: LDSC regression estimate of (SNP) heritability
 - `[["h2_se"]]`: SE of this heritability estimate.
}
\description{
Cross-Trait LD score regression
}
\keyword{internal}
