% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_and_fit.R
\name{mask_and_fit}
\alias{mask_and_fit}
\title{Mask data and Fit}
\usage{
mask_and_fit(mats, fit_function, mask_proportion = 0.05, n_times = 5,
  seed = 100)
}
\arguments{
\item{mats}{List containing beta_hat and se_hat matrices}

\item{fit_function}{Function that takes mats as input and returns E[beta_hat]}

\item{mask_proportion}{Proportion of non-missing values to mask}

\item{n_times}{Number of replications}

\item{seed}{seed}
}
\description{
Mask data and Fit
}
