% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_norm.R
\name{min_norm}
\alias{min_norm}
\title{Minimum norm distance between true and estimated factors}
\usage{
min_norm(f_true, f_hat, single_trait_thresh = 0.98, return_Q = FALSE)
}
\arguments{
\item{f_true}{Matrix of true factors (M x K1)}

\item{f_hat}{Matrix of estimated factors (M x K2)}

\item{single_trait_thresh}{Threshold to identify single trait factors. A factor is
considered a single trait factor if the maximum absolute value of its entries
is greater than this threshold after normalizing the factor to have unit norm.
Single trait factors are removed from both f_true and f_hat before matching.
Default is 0.98.}

\item{return_Q}{Logical. If TRUE, return the optimal rotation matrix.}
}
\value{
A list with the following elements:
\item{solution}{A data frame with the following columns:
\itemize{
\item true_ix: Index of the true factor
\item est_ix: Index of the matching estimated factor
\item max_true_val: Maximum absolute value of the true factor
\item max_hat_val: Maximum absolute value of the estimated factor
\item penalty: The squared Frobenius norm penalty for the matched pair
\item match_score: The matching score (absolute inner product) for the matched pair
}}
\item{frob_n}{The Frobenius norm of the difference between the best matching
factors.}
\item{Q}{(optional) The optimal matching matrix. Returned if return_Q = TRUE.}
\item{best_est}{(optional) The best matching estimated factors. Returned if return_Q = TRUE.}
\item{best_true}{(optional) The true factors corresponding to the best matching. Returned if return_Q = TRUE.}
}
\description{
Given two matrices of factors (true and estimated), find the rotation
 that minimizes the Frobenius norm of the difference between the true factors and
 rotated estimate.
}
