% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{pipeline_init}
\alias{pipeline_init}
\title{Set up Snakemake pipeline for summar statistic factors project}
\usage{
pipeline_init()
}
\description{
Set up Snakemake pipeline for summar statistic factors project
}
