% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Plot Factors from a matrix}
\usage{
plot_factors(
  x,
  row_names,
  col_names,
  row_order,
  col_order,
  row_title = "Trait",
  col_title = "Factor"
)
}
\arguments{
\item{row_names}{Labels for rows}
}
\description{
Plot Factors from a matrix
}
