% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors_bars}
\alias{plot_factors_bars}
\title{Plot factors as barplots from a matrix}
\usage{
plot_factors_bars(
  x,
  trait_names,
  factor_names,
  trait_order,
  which_factors = seq(ncol(x))
)
}
\description{
Plot factors as barplots from a matrix
}
