% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R.R
\name{R_ldsc}
\alias{R_ldsc}
\title{Calculate matrix of error correlations using LD-score regression.}
\usage{
R_ldsc(
  Z_hat,
  ldscores,
  ld_size,
  N,
  return_gencov = FALSE,
  make_well_conditioned = TRUE,
  cond_num = 1e+05,
  return_cor = TRUE
)
}
\description{
Calculate matrix of error correlations using LD-score regression.
}
