% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_ldsc.R
\name{R_ldsc_quick}
\alias{R_ldsc_quick}
\title{Cheater ld-score matrix. Fast but higher variance.}
\usage{
R_ldsc_quick(Z_hat, ldscores, weights)
}
\description{
Cheater ld-score matrix. Fast but higher variance.
}
