% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R.R
\name{R_pt}
\alias{R_pt}
\title{Calculate matrix of error correlations using p-value threshold method.}
\usage{
R_pt(
  B_hat,
  S_hat,
  p_val_thresh = 0.05,
  return_cor = TRUE,
  make_well_conditioned = TRUE,
  cond_num = 1e+05
)
}
\description{
Calculate matrix of error correlations using p-value threshold method.
}
