% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequentialGfaPrediction.R
\name{sequentialGfaPrediction}
\alias{sequentialGfaPrediction}
\title{Sequential prediction of new samples from observed data views to unobserved}
\usage{
sequentialGfaPrediction(Y, model)
}
\arguments{
\item{Y}{The new data samples, in a similar format as in function
\code{\link{gfa}}}

\item{model}{The sampled model from function \code{\link{gfa}}, with
model$opts$predict being a logical vector with the length matching the
length of Y, describing which data views will be predicted (TRUE), and
which have been observed (FALSE).}
}
\value{
A list containing the predictions, with the observed
  data sources empty. Additionally, sampling MSE is given as element 'mse',
  and likelihood as 'cost'.
}
\description{
\code{sequentialGfaPrediction} returns predictions for unobserved data
sources of a new set of data samples.
}

