% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_bh3}
\alias{sim_bh3}
\title{Simulate data from sparse factor model}
\usage{
sim_bh3(true_L, true_F, true_Theta, S, R_rows, R_cols)
}
\arguments{
\item{true_L}{Loadings matrix}

\item{true_F}{Factors matrix}

\item{S}{standard errors}

\item{R_rows}{Correlation matrix for rows of E}

\item{R_cols}{Correlation matrix for cols of E. Only one of R_row or R_cols can be given}

\item{true_Ttheta}{Theta}
}
\description{
Simulate B_hat from model B_hat = L^TF + Theta + E where elements of
E have E_ij ~ N(0, s_{ij}^2) and Cov(E_{i,.}) = SRS where R is M times M matrix of
sample correlation.
}
