% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ind_data.R
\name{sim_ind_data}
\alias{sim_ind_data}
\title{Simulate Individual Level Data}
\usage{
sim_ind_data(nind, nvar, F_mat, pi_L, pi_theta, V_theta_prop, h2, R)
}
\arguments{
\item{nind}{Number of individuals}

\item{nvar}{Number of variants}

\item{F_mat}{Trait loadings (traits by factor)}

\item{pi_theta}{Proportion of non-zero elements of theta}

\item{V_theta_prop}{Proportion of heritability contributed by theta. Length M}

\item{h2}{Total heritability. Length M}

\item{R}{covariance of environmental effects. M times M}

\item{pi_lodaings}{Proportion of non-zero elements of L}
}
\description{
Simulate Individual Level Data
}
