% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sumstats_lf.R
\name{sim_sumstats_lf}
\alias{sim_sumstats_lf}
\title{Simulate summary stats}
\usage{
sim_sumstats_lf(
  F_mat,
  N,
  J,
  h_2_trait,
  omega,
  h_2_factor,
  pi_L,
  pi_theta,
  R_E,
  R_LD,
  relative_pve,
  g_F,
  pi_F,
  overlap_prop = 1
)
}
\arguments{
\item{F_mat}{factor matrix M by K}

\item{N}{GWAS sample size}

\item{J}{Total number of SNPs to generate}

\item{h_2_trait}{Heritability of each trait. Length M vector.}

\item{omega}{Proportion of SNP heritability mediated by factors. Length M vector.}

\item{h_2_factor}{Heritability of each factor. Length K vector.}

\item{pi_L}{Proportion of non-zero elements in L_k. Length K factor}

\item{pi_theta}{Proportion of non-zero elements in theta. Scalar.}

\item{R_E}{Environmental trait correlation not mediated by factors. M by M pd matrix}

\item{g_F}{Function from which non-zero elements of F are generated}

\item{pi_F}{Propotiono of non-zero elements of F}

\item{R_LD_list}{List of eigen decompositions of LD correlation matrices, may be missing.}
}
\description{
Simulate summary statistcs for fully overlapping GWAS with no LD
}
