% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc.R
\name{snp_ldsc}
\alias{snp_ldsc}
\title{LD score regression}
\usage{
snp_ldsc(
  ld_score,
  ld_size,
  chi2,
  sample_size,
  blocks = 200,
  intercept = NULL,
  chi2_thr1 = 30,
  chi2_thr2 = Inf,
  ncores = 1,
  step1_index = NULL,
  type = c("h2", "rg"),
  w0 = NULL
)
}
\arguments{
\item{ld_score}{Vector of LD scores.}

\item{ld_size}{Number of variants used to compute `ld_score`.}

\item{chi2}{Vector of chi-squared statistics.}

\item{sample_size}{Sample size of GWAS corresponding to chi-squared statistics.
Possibly a vector, or just a single value.}

\item{blocks}{Either a single number specifying the number of blocks,
or a vector of integers specifying the block number of each `chi2` value.
Default is `200` for `snp_ldsc()`, dividing into 200 blocks of approximately
equal size. `NULL` can also be used to skip estimating standard errors,
which is the default for `snp_ldsc2()`.}

\item{intercept}{You can constrain the intercept to some value (e.g. 1).
Default is `NULL` in `snp_ldsc()` (the intercept is estimated)
and is `1` in `snp_ldsc2()` (the intercept is fixed to 1).
This is equivalent to parameter `--intercept-h2`.}

\item{chi2_thr1}{Threshold on `chi2` in step 1. Default is `30`.
This is equivalent to parameter `--two-step`.}

\item{chi2_thr2}{Threshold on `chi2` in step 2. Default is `Inf` (none).}

\item{type, w0, step1_index}{parameters used when called from snp_ldsc_rg}
}
\value{
Vector of 4 values (or only the first 2 if `blocks = NULL`):
 - `[["int"]]`: LDSC regression intercept,
 - `[["int_se"]]`: SE of this intercept,
 - `[["h2"]]`: LDSC regression estimate of (SNP) heritability (also see
   [coef_to_liab]),
 - `[["h2_se"]]`: SE of this heritability estimate.
}
\description{
LD score regression
}
\keyword{internal}
