% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sumstatFactors}
\alias{sumstatFactors}
\title{sumstatFactors}
\description{
Helper package for experimenting with factor decompostion of summary statistics.
}
\author{
Jean Morrison <jeanm@uchicago.edu>
}
