% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{undoNormalizeData}
\alias{undoNormalizeData}
\title{A function for returning predictions into the original data space}
\usage{
undoNormalizeData(pred, normalization)
}
\arguments{
\item{pred}{The predictions acquired from \code{\link{reconstruction}}.}

\item{normalization}{The output list obtained from
\code{\link{normalizeData}}.}
}
\value{
The predictions in the original data space.
}
\description{
\code{undoNormalizeData} returns the predictions on normalized data acquired
from \code{\link{normalizeData}} into the original data space.
}

