% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_L_z.R
\name{est_L_z}
\alias{est_L_z}
\title{Estimate L. This only works for z-scores using the ff method.}
\usage{
est_L_z(Z_hat, R, fit, opt1 = FALSE, opt2 = FALSE, opt3 = TRUE)
}
\arguments{
\item{Z_hat}{matrix of z-scores}

\item{R}{Estimated residual correlation of rows of Z_hat}

\item{fit}{Flash fit}
}
\description{
Estimate L. This only works for z-scores using the ff method.
}
