% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ff.R
\name{fit_ff}
\alias{fit_ff}
\title{Fit with fixed factors}
\usage{
fit_ff(
  B_hat,
  S_hat,
  N,
  R,
  kmax = 100,
  zero_thresh = 1e-15,
  adjust = TRUE,
  method = "sequential",
  max_ev_percent = 1,
  S_NULL = FALSE
)
}
\arguments{
\item{B_hat}{Matrix of (non-standardized) effect estimates SNPs by traits}

\item{S_hat}{Matrix of standard errors SNPs by traits}

\item{N}{Vector of sample sizes length equal to number of traits}

\item{R}{Estimated residual correlation matrix}

\item{kmax}{Maximum number of factors}

\item{zero_thresh}{Threshold for setting eigenvalues of R to zero}

\item{adjust}{If TRUE, will compute adjusted effect estimates as B_hat/(sqrt(N)*S_hat)}

\item{S_NULL}{If TRUE, will fit EBMF without supplying S.}
}
\value{
A list with elements fit, B_hat, L_hat, F_hat
}
\description{
Fit with fixed factors
}
