% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfa_fit.R
\name{gfa_fit}
\alias{gfa_fit}
\title{Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.}
\usage{
gfa_fit(
  Z_hat,
  B_std,
  N,
  R,
  kmax,
  ridge_penalty = 0,
  min_ev = 0.001,
  max_lr_percent = 1,
  lr_zero_thresh = 1e-10,
  max_iter = 1000,
  extrapolate = TRUE,
  ebnm_fn_F = as.ebnm.fn(prior_family = "point_normal", optmethod = "nlm"),
  ebnm_fn_L = as.ebnm.fn(prior_family = "point_normal", optmethod = "nlm"),
  init_fn = flashier::init.fn.default,
  fixed_truncate = Inf,
  duplicate_check_thresh = 0.5
)
}
\arguments{
\item{Z_hat}{A matrix of z-scores}

\item{B_std}{A matrix of standardized effects. Supply only one of B_std or Z_hat}

\item{N}{Vector of sample sizes length equal to number of traits. N is only
needed if using B_std.}

\item{R}{Estimated residual correlation matrix}

\item{kmax}{Maximum number of factors. Defaults to 2*ntraits for plain or ntraits
if using the fixed factor correction.}

\item{min_ev, max_lr_percent, lr_zero_thresh}{See details}

\item{S_inf}{A vector supplying a series of variance inflation factors for fitting ebmf.
S_inf should be decreasing and the last element should be 1.}

\item{fixed_g}{Supply the prior for loadings of fixed factors.}
}
\value{
A list with elements fit, Y, L_hat, F_hat
}
\description{
Fit with fixed factors. This version is from March 2021 and explicitly only runs on
either z-scores or standardized effects. This function replaces both fit_ff and fit_plain.
}
\details{
min_ev, max_lr_percent, lr_zero_thresh: We are approximating the matrix R as VDV^T + lambda I
where lambda is the smallest eigenvalue of R. min_ev specifies the smallest acceptable value of lambda.
max_lr_percent specifies the proportion of variance of (R - lambda I) that will be retained. This defaults
to 1 making the approximation exact. Eigenvalues of (R-lambda I) that are less than lr_zero_thresh will
be set to zero.
}
