% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings_gls.R
\name{gfa_loadings_gls}
\alias{gfa_loadings_gls}
\title{GFA generalized least squares loadings}
\usage{
gfa_loadings_gls(beta_hat, S, fit)
}
\arguments{
\item{beta_hat}{Variants by traits matrix of association estimates. Traits
should be in the same order as used to produce fitted gfa object. The variant set may be different than
the variant set used to produce the fitted gfa object.}

\item{S}{Variants by traits matrix of standard errors. Variants and traits should match variants and traits in beta_hat.}

\item{fit}{Object produced by gfa_ft().}
}
\value{
A list with elements L (estimated loadings), S (standard errors for loadings), and P (p-values). All variant by factor matrices.
}
\description{
Compute GLS estimates and p-values for loadings
}
