% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_ldsc.R
\name{R_ldsc}
\alias{R_ldsc}
\title{Calculate matrix of error correlations using LD-score regression.}
\usage{
R_ldsc(Z_hat, ldscores, ld_size, N, return_gencov = FALSE)
}
\description{
Calculate matrix of error correlations using LD-score regression.
}
