% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_ldsc.R
\name{R_pt}
\alias{R_pt}
\title{Calculate matrix of error correlations using p-value threshold method.}
\usage{
R_pt(B_hat, S_hat, N, p_val_thresh = 0.05, z_scores = FALSE)
}
\description{
Calculate matrix of error correlations using p-value threshold method.
}
