% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R.R
\name{R_pt}
\alias{R_pt}
\title{Calculate matrix of error correlations using p-value threshold method.}
\usage{
R_pt(
  B_hat,
  S_hat,
  p_val_thresh = 0.05,
  return_cor = TRUE,
  make_well_conditioned = TRUE,
  cond_num = 1e+05
)
}
\arguments{
\item{B_hat}{Matrix of effect size estimates}

\item{S_hat}{Matrix of standard errors}

\item{p_val_thresh}{P-value threshold}

\item{return_cor}{Return matrix as correlation matrix (apply cov2cor), recommended}

\item{cond_num}{Condition number if using make_well_conditioned}

\item{max_well_conditioned}{Project matrix to nearest well conditioned positive definite
matrix using Matrix::nearPD}
}
\value{
A traits by traits matrix of nuisance correlations.
}
\description{
Calculate matrix of error correlations using p-value threshold method.
}
